refactor(electron): typescript check on build scripts
