+++ apps/electron/scripts/build-layers.ts
@@ -1,23 +1,12 @@
-#!/usr/bin/env zx
-import 'zx/globals';
-
 import * as esbuild from 'esbuild';
 
-import { config } from './common.mjs';
-
-const NODE_ENV =
-  process.env.NODE_ENV === 'development' ? 'development' : 'production';
-
-if (process.platform === 'win32') {
-  $.shell = true;
-  $.prefix = '';
-}
+import { config, mode } from './common';
 
 async function buildLayers() {
   const common = config();
 
-  const define = {
-    'process.env.NODE_ENV': `"${NODE_ENV}"`,
+  const define: Record<string, string> = {
+    'process.env.NODE_ENV': `"${mode}"`,
     'process.env.BUILD_TYPE': `"${process.env.BUILD_TYPE || 'stable'}"`,
   };
 
@@ -28,10 +17,10 @@ async function buildLayers() {
   }
 
   await esbuild.build({
-    ...common.layers,
+    ...common,
     define: define,
   });
 }
 
 await buildLayers();
-echo('Build layers done');
+console.log('Build layers done');
