+++ apps/electron/tsconfig.node.json
@@ -9,7 +9,16 @@
     "allowSyntheticDefaultImports": true,
     "noEmit": false,
     "outDir": "./lib/scripts",
+    "types": ["node"],
     "allowJs": true
   },
-  "include": ["./scripts", "esbuild.main.config.ts", "esbuild.plugin.config.ts"]
+  "include": [
+    "./scripts",
+    "esbuild.main.config.ts",
+    "esbuild.plugin.config.ts"
+  ],
+  "ts-node": {
+    "esm": true,
+    "experimentalSpecifierResolution": "node"
+  }
 }
